# wasi: optimize and benchmark fd_read