--- a/imports/wasi_snapshot_preview1/fs.go
+++ b/imports/wasi_snapshot_preview1/fs.go
@@ -613,6 +613,7 @@ func fdReadFn(ctx context.Context, mod api.Module, stack []uint64) (nread uint32
 func fdReadOrPread(ctx context.Context, mod api.Module, stack []uint64, isPread bool) (uint32, Errno) {
 	sysCtx := mod.(*wasm.CallContext).Sys
 	mem := mod.Memory()
+
 	fd := uint32(stack[0])
 	iovs := uint32(stack[1])
 	iovsCount := uint32(stack[2])
@@ -639,7 +640,7 @@ func fdReadOrPread(ctx context.Context, mod api.Module, stack []uint64, isPread 
 
 	var nread uint32
 	iovsStop := iovsCount << 3 // iovsCount * 8
-	iovsBuf, ok := mod.Memory().Read(ctx, iovs, iovsStop)
+	iovsBuf, ok := mem.Read(ctx, iovs, iovsStop)
 	if !ok {
 		return 0, ErrnoFault
 	}
@@ -1102,11 +1103,13 @@ var fdWrite = proxyResultParams(&wasm.HostFunc{
 }, fdWriteName)
 
 func fdWriteFn(ctx context.Context, mod api.Module, stack []uint64) (uint32, Errno) {
+	sysCtx := mod.(*wasm.CallContext).Sys
+	mem := mod.Memory()
+
 	fd := uint32(stack[0])
 	iovs := uint32(stack[1])
 	iovsCount := uint32(stack[2])
 
-	sysCtx := mod.(*wasm.CallContext).Sys
 	writer := internalsys.FdWriter(ctx, sysCtx, fd)
 	if writer == nil {
 		return 0, ErrnoBadf
@@ -1115,7 +1118,7 @@ func fdWriteFn(ctx context.Context, mod api.Module, stack []uint64) (uint32, Err
 	var err error
 	var nwritten uint32
 	iovsStop := iovsCount << 3 // iovsCount * 8
-	iovsBuf, ok := mod.Memory().Read(ctx, iovs, iovsStop)
+	iovsBuf, ok := mem.Read(ctx, iovs, iovsStop)
 	if !ok {
 		return 0, ErrnoFault
 	}
@@ -1128,7 +1131,7 @@ func fdWriteFn(ctx context.Context, mod api.Module, stack []uint64) (uint32, Err
 		if writer == io.Discard { // special-case default
 			n = int(l)
 		} else {
-			b, ok := mod.Memory().Read(ctx, offset, l)
+			b, ok := mem.Read(ctx, offset, l)
 			if !ok {
 				return 0, ErrnoFault
 			}