```bash
# before change
$ go test -run='^$' -bench '^Benchmark_fdRead$' -count=5 > old.txt
# after change
$ go test -run='^$' -bench '^Benchmark_fdRead$' -count=5 > new.txt
$ benchstat old.txt new.txt
name             old time/op  new time/op  delta
_fdRead/1x8-16    157ns ± 5%   133ns ± 1%  -15.34%  (p=0.008 n=5+5)
_fdRead/2x16-16   193ns ± 6%   155ns ± 1%  -19.77%  (p=0.008 n=5+5)
```